tests/wpt/css-tests/cssom-view-1_dev/xhtml1/cssom-getBoundingClientRect-002.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <meta charset="utf-8" />
    <title>CSS Test (CSSOM View): getBoundingClientRect of element outside DOM</title>
    <link href="http://chrisrebert.com" rel="author" title="Chris Rebert" />
    <link href="http://www.w3.org/TR/cssom-view/#dom-element-getclientrects" rel="help" />
    <link href="http://www.w3.org/TR/cssom-view/#dom-element-getboundingclientrect" rel="help" />
    <meta content="dom" name="flags" />
    <meta content="Calling getBoundingClientRect on an element that is outside of the DOM (and therefore does not have an associated layout box) should result in an all-zeroes DOMRect and should definitely not throw an error." name="assert" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        test(
            function () {
                var rect = document.createElement('div').getBoundingClientRect();
                assert_equals(rect.x, 0, "DOMRect's x should be zero");
                assert_equals(rect.y, 0, "DOMRect's y should be zero");
                assert_equals(rect.width, 0, "DOMRect's width should be zero");
                assert_equals(rect.height, 0, "DOMRect's height should be zero");
            },
            "getBoundingClientRect on a newly-created Element not yet inserted into the DOM should return an all-zeroes DOMRect"
        );
    </script>
</head>
<body>

</body></html>
